Second Commit - Image Upload Issue Solved

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -38,18 +38,21 @@ public class ImageController {
     //This method is called when the details of the specific image with corresponding title are to be displayed
     //The logic is to get the image from the databse with corresponding title. After getting the image from the database the details are shown
     //First receive the dynamic parameter in the incoming request URL in a string variable 'title' and also the Model type object
-    //Call the getImageByTitle() method in the business logic to fetch all the details of that image
+    //Call the getImage(imageId) method in the business logic to fetch all the details of that image
     //Add the image in the Model type object with 'image' as the key
     //Return 'images/image.html' file
 
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+    @RequestMapping("/images/{imageId}/{title}")
+    public String showImage(@PathVariable(name = "imageId") Integer imageId,
+                            @PathVariable(name ="title") String title, Model model) {
+        Image image = imageService.getImage(imageId);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+        //Adding comment attribute
+//        model.addAttribute("comments",image.getComments());
         return "images/image";
     }
 
@@ -92,13 +95,10 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
-
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-        String error = "Only the owner of the image can edit the image";
-        String tags = convertTagsToString(image.getTags());
 
-        model.addAttribute("editError", error);
+        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
         return "images/edit";
@@ -135,7 +135,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + image.getId() + "/" + updatedImage.getTitle();
     }
 
 
@@ -143,13 +143,26 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,
+                                    Model model,
+                                    HttpSession session) {
+        Image image = imageService.getImage(imageId);
 
-        String error = "Only the owner of the image can edit the image";
-        model.addAttribute("deleteError", error);
+        User user = (User) session.getAttribute("loggeduser");
 
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+        //Added Validation for user id for owner to delete the image
+        if(image.getUser().getId() != user.getId()){
+            String error = "Only the owner of the image can delete the image";
+
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }else
+        {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
